+++ crates/dock-alloc-solver/src/domain.rs
@@ -19,10 +19,9 @@
 // OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-use std::fmt::Display;
-
-use dock_alloc_core::domain::{SpaceInterval, TimeInterval};
+use dock_alloc_core::{space::SpaceInterval, time::TimeInterval};
 use num_traits::{PrimInt, Signed};
+use std::fmt::Display;
 
 /// A version number used to track iterations or generations in optimization algorithms.
 ///
@@ -177,7 +176,7 @@ impl<T: PrimInt + Signed> From<(SpaceInterval, TimeInterval<T>)> for SpaceTimeRe
 #[cfg(test)]
 mod tests {
     use super::*;
-    use dock_alloc_core::domain::{SpacePosition, TimePoint};
+    use dock_alloc_core::{space::SpacePosition, time::TimePoint};
 
     #[test]
     fn test_space_time_rectangle_creation() {
